+++ lending/CodatLending/LoanWritebackPayments.cs
@@ -56,17 +56,17 @@ namespace CodatLending
         /// 
         /// </remarks>
         /// </summary>
-        Task<GetCreatePaymentsModelResponse> GetCreateModelAsync(GetCreatePaymentsModelRequest? request = null);
+        Task<GetCreatePaymentModelResponse> GetCreateModelAsync(GetCreatePaymentModelRequest? request = null);
     }
 
     public class LoanWritebackPaymentsSDK: ILoanWritebackPaymentsSDK
     {
         public SDKConfig Config { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "4.3.0";
-        private const string _sdkGenVersion = "2.159.2";
+        private const string _sdkVersion = "4.4.0";
+        private const string _sdkGenVersion = "2.173.0";
         private const string _openapiDocVersion = "3.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 4.3.0 2.159.2 3.0.0 Codat.Lending";
+        private const string _userAgent = "speakeasy-sdk/csharp 4.4.0 2.173.0 3.0.0 Codat.Lending";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -121,7 +121,7 @@ namespace CodatLending
                 
                 return response;
             }
-            if((response.StatusCode == 400) || (response.StatusCode == 401) || (response.StatusCode == 404) || (response.StatusCode == 429))
+            if((response.StatusCode == 400) || (response.StatusCode == 401) || (response.StatusCode == 402) || (response.StatusCode == 403) || (response.StatusCode == 404) || (response.StatusCode == 429) || (response.StatusCode == 500) || (response.StatusCode == 503))
             {
                 if(Utilities.IsContentTypeMatch("application/json",response.ContentType))
                 {
@@ -134,7 +134,7 @@ namespace CodatLending
         }
         
 
-        public async Task<GetCreatePaymentsModelResponse> GetCreateModelAsync(GetCreatePaymentsModelRequest? request = null)
+        public async Task<GetCreatePaymentModelResponse> GetCreateModelAsync(GetCreatePaymentModelRequest? request = null)
         {
             string baseUrl = _serverUrl;
             if (baseUrl.EndsWith("/"))
@@ -154,7 +154,7 @@ namespace CodatLending
 
             var contentType = httpResponse.Content.Headers.ContentType?.MediaType;
             
-            var response = new GetCreatePaymentsModelResponse
+            var response = new GetCreatePaymentModelResponse
             {
                 StatusCode = (int)httpResponse.StatusCode,
                 ContentType = contentType,
@@ -170,7 +170,7 @@ namespace CodatLending
                 
                 return response;
             }
-            if((response.StatusCode == 401) || (response.StatusCode == 404) || (response.StatusCode == 429))
+            if((response.StatusCode == 401) || (response.StatusCode == 402) || (response.StatusCode == 403) || (response.StatusCode == 404) || (response.StatusCode == 429) || (response.StatusCode == 500) || (response.StatusCode == 503))
             {
                 if(Utilities.IsContentTypeMatch("application/json",response.ContentType))
                 {