+++ lending/CodatLending/Models/Operations/DownloadAccountingInvoicePdfResponse.cs
@@ -10,6 +10,7 @@
 #nullable enable
 namespace CodatLending.Models.Operations
 {
+    using CodatLending.Models.Shared;
     using System.Net.Http;
     using System;
     
@@ -25,6 +26,11 @@ namespace CodatLending.Models.Operations
         /// Success
         /// </summary>
         public byte[]? Data { get; set; }
+
+        /// <summary>
+        /// Your API request was not properly authorized.
+        /// </summary>
+        public ErrorMessage? ErrorMessage { get; set; }
 
         /// <summary>
         /// HTTP response status code for this operation