+++ lending/CodatLending/Models/Operations/GetCategorizedBankStatementResponse.cs
@@ -28,7 +28,7 @@ namespace CodatLending.Models.Operations
         public EnhancedCashFlowTransactions? EnhancedCashFlowTransactions { get; set; }
 
         /// <summary>
-        /// Your API request was not properly authorized.
+        /// Your `query` parameter was not correctly formed
         /// </summary>
         public ErrorMessage? ErrorMessage { get; set; }