+++ lending/CodatLending/Models/Operations/GetExcelReportGenerationStatusResponse.cs
@@ -23,7 +23,7 @@ namespace CodatLending.Models.Operations
         public string? ContentType { get; set; } = default!;
 
         /// <summary>
-        /// Your API request was not properly authorized.
+        /// Your `query` parameter was not correctly formed
         /// </summary>
         public ErrorMessage? ErrorMessage { get; set; }