+++ lending/RELEASES.md
@@ -153,3 +153,13 @@ Based on:
 - [csharp v4.3.0] lending
 ### Releases
 - [NuGet v4.3.0] https://www.nuget.org/packages/Codat.Lending/4.3.0 - lending
+
+## 2023-11-06 01:27:35
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.4.0] lending
+### Releases
+- [NuGet v4.4.0] https://www.nuget.org/packages/Codat.Lending/4.4.0 - lending